# swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Projeto Vale - Sistema de Gestão Imobiliária",
    "description": "API completa para o sistema de gestão de lotes imobiliários Projeto Vale. Inclui gerenciamento de lotes, clientes, transações, autenticação e relatórios.",
    "version": "1.0.0",
    "contact": {
      "name": "Projeto Vale",
      "email": "contato@projeto-vale.com.br",
      "url": "https://projeto-vale.com.br"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "Servidor de desenvolvimento"
    },
    {
      "url": "https://api.projeto-vale.com.br",
      "description": "Servidor de produção"
    }
  ],
  "security": [
    {
      "BearerAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Token JWT obtido através do endpoint /auth/login"
      }
    },
    "schemas": {
...
      "Lot": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "cuid",
            "description": "ID único do lote"
          },
          "empreendimento": {
            "type": "string",
            "description": "ID ou nome do empreendimento"
          },
          "quadra": {
            "type": "string",
            "description": "Identificador da quadra"
          },
          "lote": {
            "type": "string",
            "description": "Identificador do lote"
          },
          "area": {
            "type": "number",
            "format": "float",
            "description": "Área do lote em metros quadrados"
          },
          "valor": {
            "type": "number",
            "format": "float",
            "description": "Valor total do lote"
          },
          "entrada": {
            "type": "number",
            "format": "float",
            "description": "Valor de entrada mínimo"
          },
          "status": {
            "type": "string",
            "enum": ["DISPONÍVEL", "RESERVADO", "EM PROPOSTA", "VENDIDO"],
            "description": "Status atual do lote"
          },
          "observacoes": {
            "type": "string",
            "description": "Observações sobre o lote"
          },
          "reservedAt": {
            "type": "string",
            "format": "date-time",
            "description": "Data da reserva (se aplicável)"
          },
          "reservedBy": {
            "type": "string",
            "description": "ID do usuário que reservou o lote"
          },
          "managedBy": {
            "type": "string",
            "description": "ID do administrador que gerencia o lote"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "Data de criação do lote"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "description": "Data da última atualização"
          }
        },
        "required": ["empreendimento", "quadra", "lote", "area", "valor", "entrada", "status"]
      },
      "CreateLotRequest": {
        "type": "object",
        "properties": {
          "empreendimento": {
            "type": "string",
            "description": "ID ou nome do empreendimento"
          },
          "quadra": {
            "type": "string",
            "description": "Identificador da quadra"
          },
          "lote": {
            "type": "string",
            "description": "Identificador do lote"
          },
          "area": {
            "type": "number",
            "format": "float",
            "description": "Área do lote em metros quadrados"
          },
          "valor": {
            "type": "number",
            "format": "float",
            "description": "Valor total do lote"
          },
          "entrada": {
            "type": "number",
            "format": "float",
            "description": "Valor de entrada mínimo"
          },
          "status": {
            "type": "string",
            "enum": ["DISPONÍVEL", "RESERVADO", "EM PROPOSTA", "VENDIDO"],
            "default": "DISPONÍVEL",
            "description": "Status inicial do lote"
          },
          "observacoes": {
            "type": "string",
            "description": "Observações sobre o lote"
          }
        },
        "required": ["empreendimento", "quadra", "lote", "area", "valor", "entrada", "status"]
      },
      "UpdateLotRequest": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "ID do lote a ser atualizado"
          },
          "empreendimento": {
            "type": "string",
            "description": "ID ou nome do empreendimento"
          },
          "quadra": {
            "type": "string",
            "description": "Identificador da quadra"
          },
          "lote": {
            "type": "string",
            "description": "Identificador do lote"
          },
          "area": {
            "type": "number",
            "format": "float",
            "description": "Área do lote em metros quadrados"
          },
          "valor": {
            "type": "number",
            "format": "float",
            "description": "Valor total do lote"
          },
          "entrada": {
            "type": "number",
            "format": "float",
            "description": "Valor de entrada mínimo"
          },
          "status": {
            "type": "string",
            "enum": ["DISPONÍVEL", "RESERVADO", "EM PROPOSTA", "VENDIDO"],
            "description": "Novo status do lote"
          },
          "observacoes": {
            "type": "string",
            "description": "Observações sobre o lote"
          },
          "reservedBy": {
            "type": "string",
            "description": "ID do usuário que reservou o lote"
          }
        },
        "required": ["id"]
      },
      "LotsResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Lot"
            }
          },
          "metadata": {
            "type": "object",
            "properties": {
              "total": {
                "type": "integer",
                "description": "Total de lotes retornados"
              },
              "filters": {
                "type": "object",
                "properties": {
                  "status": {
                    "type": "string",
                    "description": "Filtro de status aplicado"
                  },
                  "empreendimento": {
                    "type": "string",
                    "description": "Filtro de empreendimento aplicado"
                  },
                  "quadra": {
                    "type": "string",
                    "description": "Filtro de quadra aplicado"
                  },
                  "lote": {
                    "type": "string",
                    "description": "Filtro de lote aplicado"
                  }
                }
              },
              "timestamp": {
                "type": "string",
                "format": "date-time",
                "description": "Timestamp da consulta"
              }
            }
          }
        },
        "required": ["data", "metadata"]
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "Mensagem de erro"
          },
          "code": {
            "type": "integer",
            "description": "Código do erro HTTP"
          },
          "details": {
            "type": "object",
            "description": "Detalhes adicionais do erro"
          }
        },
        "required": ["error"]
      },
...
    }
  },
  "paths": {
...
    "/lots": {
      "get": {
        "tags": ["Lotes"],
        "summary": "Listar lotes",
        "description": "Retorna uma lista de lotes com suporte a filtros",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "description": "Filtrar por status do lote",
            "schema": {
              "type": "string",
              "enum": ["DISPONÍVEL", "RESERVADO", "EM PROPOSTA", "VENDIDO"]
            }
          },
          {
            "name": "empreendimento",
            "in": "query",
            "description": "Filtrar por empreendimento",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "quadra",
            "in": "query",
            "description": "Filtrar por quadra",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "lote",
            "in": "query",
            "description": "Filtrar por número do lote",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "Número da página (paginação)",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Limite de resultados por página",
            "schema": {
              "type": "integer",
              "default": 50,
              "maximum": 1000
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de lotes retornada com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LotsResponse"
                }
              }
            }
          },
          "400": {
            "description": "Parâmetros inválidos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Erro interno do servidor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Lotes"],
        "summary": "Criar novo lote",
        "description": "Cria um novo lote no sistema",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateLotRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Lote criado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/Lot"
                    },
                    "message": {
                      "type": "string",
                      "description": "Mensagem de sucesso"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Dados inválidos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Lote já existe",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Erro interno do servidor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["Lotes"],
        "summary": "Atualizar lote existente",
        "description": "Atualiza os dados de um lote existente",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateLotRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Lote atualizado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/Lot"
                    },
                    "message": {
                      "type": "string",
                      "description": "Mensagem de sucesso"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Dados inválidos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Lote não encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Erro interno do servidor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Lotes"],
        "summary": "Excluir lote",
        "description": "Exclui um lote do sistema",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "required": true,
            "description": "ID do lote a ser excluído",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lote excluído com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "description": "Mensagem de sucesso"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "ID não fornecido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Lote não encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Erro interno do servidor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}